Be a bit more frugal with our compression buffer
For small files (and the remaining bit of large ones, though this is not
as big of a deal), the compression buffer can be sized based on the
remaining data rather than the block size. This often allows it to avoid
being classified as a "medium sized" allocation, saving an expensive
call to madvise when the buffer is freed.