Remove warning about using unxip in your program
This is obviously at odds to vending libunxip. More specifically,
though, while unxip schedules all over the cooperative thread pool, most
clients seem to not really expect to use it heavily anyways (since they
are small wrappers around the API, perhaps with some UI). For those that
actually do plan to do work there, Swift Concurrency does an OK job of
having the work share available resources. I don't expect this to be a
major problem and those that it is one for can probably figure out how
to work around it themselves, or reach out and ask in the worst case.