Restructure DataReader to submit chunks in order
I can't remember who reported this but if you are using unxip as a
library it can surface a bug because we submit to the stream
asynchronously without waiting for it to be enqueued, which lets the
chunks be vended out of order. This does not reproduce if you build
unxip itself, I assume because the compiler doesn't actually hop
executors in that case. We don't need this async submission anyway and I
think the fixed code is more straightforward.